# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Control de Capital Humano API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  superadmin: 'superadmin',
  admin: 'admin',
  user: 'user',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  puesto: string;
  area: string;
  /** @nullable */
  profile_id?: number | null;
  /** @nullable */
  udn_id: number | null;
  role: UserRole;
  created_at: string;
}

export interface AuthResponse {
  user: User;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  superadmin: 'superadmin',
  admin: 'admin',
  user: 'user',
} as const;

export interface UserInput {
  name: string;
  email: string;
  password: string;
  puesto: string;
  area: string;
  /** @nullable */
  profile_id?: number | null;
  /** @nullable */
  udn_id?: number | null;
  role?: UserInputRole;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  superadmin: 'superadmin',
  admin: 'admin',
  user: 'user',
} as const;

export interface UserUpdate {
  name?: string;
  email?: string;
  password?: string;
  puesto?: string;
  area?: string;
  /** @nullable */
  profile_id?: number | null;
  /** @nullable */
  udn_id?: number | null;
  role?: UserUpdateRole;
}

export interface Udn {
  id: number;
  name: string;
  code: string;
  /** @nullable */
  description?: string | null;
  created_at: string;
}

export interface UdnInput {
  name: string;
  code: string;
  description?: string;
}

export interface UdnUpdate {
  name?: string;
  code?: string;
  description?: string;
}

export interface Profile {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  permissions: string[];
  created_at: string;
}

export interface ProfileInput {
  name: string;
  description?: string;
  permissions: string[];
}

export interface ProfileUpdate {
  name?: string;
  description?: string;
  permissions?: string[];
}

export interface Zone {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  udn_id?: number | null;
  created_at: string;
}

export interface ZoneInput {
  name: string;
  description?: string;
  udn_id?: number;
}

export interface ZoneUpdate {
  name?: string;
  description?: string;
  udn_id?: number;
}

export interface Panel {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  diagram_url?: string | null;
  /** Number of columns (numbered) */
  columns: number;
  /** Number of rows (lettered A, B, C...) */
  rows: number;
  /** @nullable */
  zone_id?: number | null;
  /** @nullable */
  side_id?: number | null;
  /** @nullable */
  visual_zone_id?: number | null;
  alphanumeric_ids?: number[];
  created_at: string;
}

export interface PanelInput {
  name: string;
  description?: string;
  diagram_url?: string;
  columns: number;
  rows: number;
  zone_id?: number;
  side_id?: number;
  visual_zone_id?: number;
  alphanumeric_ids?: number[];
}

export interface PanelUpdate {
  name?: string;
  description?: string;
  diagram_url?: string;
  columns?: number;
  rows?: number;
  zone_id?: number;
  side_id?: number;
  visual_zone_id?: number;
  alphanumeric_ids?: number[];
}

export interface AuditCapture {
  id: number;
  unit_number: number;
  week_number: number;
  date: string;
  skill_number: string;
  /** @nullable */
  zone_id?: number | null;
  /** @nullable */
  panel_id?: number | null;
  /** @nullable */
  side_id?: number | null;
  /** @nullable */
  visual_zone_id?: number | null;
  /** @nullable */
  alphanumeric_id?: number | null;
  grid_col: number;
  grid_row: string;
  /** @nullable */
  defect_id?: number | null;
  /** @nullable */
  defect_other?: string | null;
  quantity: number;
  created_at: string;
}

export interface AuditCaptureInput {
  unit_number: number;
  week_number: number;
  date: string;
  skill_number: string;
  zone_id?: number;
  panel_id?: number;
  side_id?: number;
  visual_zone_id?: number;
  alphanumeric_id?: number;
  grid_col: number;
  grid_row: string;
  defect_id?: number;
  defect_other?: string;
  quantity: number;
}

export interface AuditCaptureUpdate {
  skill_number?: string;
  zone_id?: number;
  panel_id?: number;
  side_id?: number;
  visual_zone_id?: number;
  alphanumeric_id?: number;
  grid_col?: number;
  grid_row?: string;
  defect_id?: number;
  defect_other?: string;
  quantity?: number;
}

export interface AuditDailyCounter {
  date: string;
  next_unit_number: number;
  week_number: number;
}

export interface Defect {
  id: number;
  name: string;
  code: string;
  /** @nullable */
  description?: string | null;
  created_at: string;
}

export interface DefectInput {
  name: string;
  code: string;
  description?: string;
}

export interface DefectUpdate {
  name?: string;
  code?: string;
  description?: string;
}

export interface Side {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  created_at: string;
}

export interface SideInput {
  name: string;
  description?: string;
}

export interface SideUpdate {
  name?: string;
  description?: string;
}

export interface VisualZone {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  created_at: string;
}

export interface VisualZoneInput {
  name: string;
  description?: string;
}

export interface VisualZoneUpdate {
  name?: string;
  description?: string;
}

export interface AlphanumericRecord {
  id: number;
  name: string;
  code: string;
  /** @nullable */
  description?: string | null;
  created_at: string;
}

export interface AlphanumericInput {
  name: string;
  code: string;
  description?: string;
}

export interface AlphanumericUpdate {
  name?: string;
  code?: string;
  description?: string;
}

export interface DashboardStats {
  totalUsers: number;
  totalUdns: number;
  totalZones: number;
  totalDefects: number;
  totalPanels: number;
  recentUsers: User[];
}

export type ListAuditCapturesParams = {
/**
 * Filter by date (YYYY-MM-DD)
 */
date?: string;
panel_id?: number;
};

export type GetAuditDailyCounterParams = {
date: string;
};