# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Control de Capital Humano API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # USERS
...
  /defects:
    get:
      operationId: listDefects
      tags: [defects]
      summary: List all defects
      responses:
        "200":
          description: List of defects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Defect"
    post:
      operationId: createDefect
      tags: [defects]
      summary: Create a defect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DefectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Defect"
  /defects/{id}:
    get:
      operationId: getDefect
      tags: [defects]
      summary: Get defect by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Defect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Defect"
    patch:
      operationId: updateDefect
      tags: [defects]
      summary: Update defect
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DefectUpdate"
      responses:
        "200":
          description: Updated defect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Defect"
    delete:
      operationId: deleteDefect
      tags: [defects]
      summary: Delete defect
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # LADOS
...
  /audit-captures:
    get:
      operationId: listAuditCaptures
      tags: [auditCaptures]
      summary: List audit captures
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Filter by date (YYYY-MM-DD)
        - name: panel_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of audit captures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditCapture"
    post:
      operationId: createAuditCapture
      tags: [auditCaptures]
      summary: Create an audit capture record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditCaptureInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditCapture"
  /audit-captures/daily-counter:
    get:
      operationId: getAuditDailyCounter
      tags: [auditCaptures]
      summary: Get next unit number for a given date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily counter info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditDailyCounter"
  /audit-captures/{id}:
    get:
      operationId: getAuditCapture
      tags: [auditCaptures]
      summary: Get audit capture by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AuditCapture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditCapture"
        "404":
          description: Not found
    patch:
      operationId: updateAuditCapture
      tags: [auditCaptures]
      summary: Update audit capture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditCaptureUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditCapture"
    delete:
      operationId: deleteAuditCapture
      tags: [auditCaptures]
      summary: Delete audit capture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ALFANUMERICO
  /alphanumeric:
    get:
      operationId: listAlphanumeric
      tags: [alphanumeric]
      summary: List all alphanumeric records
      responses:
        "200":
          description: List of alphanumeric records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlphanumericRecord"
    post:
      operationId: createAlphanumeric
      tags: [alphanumeric]
      summary: Create an alphanumeric record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlphanumericInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlphanumericRecord"
  /alphanumeric/{id}:
    get:
      operationId: getAlphanumeric
      tags: [alphanumeric]
      summary: Get alphanumeric record by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AlphanumericRecord
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlphanumericRecord"
    patch:
      operationId: updateAlphanumeric
      tags: [alphanumeric]
      summary: Update alphanumeric record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlphanumericUpdate"
      responses:
        "200":
          description: Updated alphanumeric record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlphanumericRecord"
    delete:
      operationId: deleteAlphanumeric
      tags: [alphanumeric]
      summary: Delete alphanumeric record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # DASHBOARD
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # USER
    User:
      type: object
      required: [id, name, email, puesto, area, udn_id, role, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        puesto:
          type: string
        area:
          type: string
        profile_id:
          type: ["integer", "null"]
        udn_id:
          type: ["integer", "null"]
        role:
          type: string
          enum: [superadmin, admin, user]
        created_at:
          type: string
          format: date-time
...
    AuditCapture:
      type: object
      required: [id, unit_number, week_number, date, skill_number, grid_col, grid_row, quantity, created_at]
      properties:
        id:
          type: integer
        unit_number:
          type: integer
        week_number:
          type: integer
        date:
          type: string
          format: date
        skill_number:
          type: string
        zone_id:
          type: ["integer", "null"]
        panel_id:
          type: ["integer", "null"]
        side_id:
          type: ["integer", "null"]
        visual_zone_id:
          type: ["integer", "null"]
        alphanumeric_id:
          type: ["integer", "null"]
        grid_col:
          type: integer
        grid_row:
          type: string
        defect_id:
          type: ["integer", "null"]
        defect_other:
          type: ["string", "null"]
        quantity:
          type: integer
        created_at:
          type: string
          format: date-time
    AuditCaptureInput:
      type: object
      required: [unit_number, week_number, date, skill_number, grid_col, grid_row, quantity]
      properties:
        unit_number:
          type: integer
        week_number:
          type: integer
        date:
          type: string
          format: date
        skill_number:
          type: string
        zone_id:
          type: integer
        panel_id:
          type: integer
        side_id:
          type: integer
        visual_zone_id:
          type: integer
        alphanumeric_id:
          type: integer
        grid_col:
          type: integer
        grid_row:
          type: string
        defect_id:
          type: integer
        defect_other:
          type: string
        quantity:
          type: integer
    AuditCaptureUpdate:
      type: object
      properties:
        skill_number:
          type: string
        zone_id:
          type: integer
        panel_id:
          type: integer
        side_id:
          type: integer
        visual_zone_id:
          type: integer
        alphanumeric_id:
          type: integer
        grid_col:
          type: integer
        grid_row:
          type: string
        defect_id:
          type: integer
        defect_other:
          type: string
        quantity:
          type: integer
    AuditDailyCounter:
      type: object
      required: [date, next_unit_number, week_number]
      properties:
        date:
          type: string
          format: date
        next_unit_number:
          type: integer
        week_number:
          type: integer

    # DEFECT
    Defect:
      type: object
      required: [id, name, code, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
    DefectInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
    DefectUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string

    # SIDE
...
    AlphanumericRecord:
      type: object
      required: [id, name, code, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
    AlphanumericInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
    AlphanumericUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string

    # DASHBOARD
...