# lib/api-zod/src/generated/types/auditCapture.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Control de Capital Humano API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditCapture {
  id: number;
  unit_number: number;
  week_number: number;
  date: Date;
  skill_number: string;
  /** @nullable */
  zone_id?: number | null;
  /** @nullable */
  panel_id?: number | null;
  /** @nullable */
  side_id?: number | null;
  /** @nullable */
  visual_zone_id?: number | null;
  /** @nullable */
  alphanumeric_id?: number | null;
  grid_col: number;
  grid_row: string;
  /** @nullable */
  defect_id?: number | null;
  /** @nullable */
  defect_other?: string | null;
  quantity: number;
  created_at: Date;
}