# lib/api-zod/src/generated/types/auditCaptureUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Control de Capital Humano API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditCaptureUpdate {
  skill_number?: string;
  zone_id?: number;
  panel_id?: number;
  side_id?: number;
  visual_zone_id?: number;
  alphanumeric_id?: number;
  grid_col?: number;
  grid_row?: string;
  defect_id?: number;
  defect_other?: string;
  quantity?: number;
}